Text plug-ins guess whether administrator rights are needed to clean an item. If so TextPlugInfo->AdminPriviligesRequired() returns true.
DC automatically hides/skips plugins that need admin rights when the user is not an administrator. 

Added context menu for 'Select All', 'Select None' and 'Invert Selection'

Added hidden button for saving cleaning results to file. Not yet implemented.

Added hidden checkbox to forcibly show admin-level plugins when the user is not an admin. Not yet implemented.